Added option to preserve mediaqueries without needing a separate style tag
The current approach requires media queries are kept in a separate style
tag, however this is not possible to do with many frameworks or CSS
preprocessors.

I've added a new option for keeping media queries that results ina new
style tag being added at the top of the body (important for many
clients) with the extracted mediaqueries. This only runs if
removeStyleElements is true to avoid duplication.

Thereis one breaking changes with how mediaqueries are currently
handled:

Allowed media queries in stylesheets are no longer parsed. Media query
support in emails is poor anyway so screen only really shouldn't be
used. If it is used it is likely to be explicitly to target certain
clients that only read media queries (this does not affect media
attributes)

I believe this change is worthwhile for the benefit gained in allowing
Premailer.Net to work with various preprocessors and is a minor edge
case in general if you are designing for email properly. Updating CSS to
avoid these changes will likely actually improve email design support.